Change code structure
