+++ App/src/test/java/objects/Account.java
@@ -1,0 +1,14 @@
+package objects;
+
+/**
+ * @author tien.hoang
+ */
+public class Account {
+    public String email;
+    public String password;
+
+    public Account(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+}
