+++ Core/src/main/java/at/base/Constant.java
@@ -1,0 +1,9 @@
+package at.base;
+
+final class Constant {
+    static final int DEFAULT_TIME_OUT = 10;
+
+    private Constant() {
+        // No-op
+    }
+}
