+++ Core/src/main/java/at/core/PropertyListener.java
