+++ Core/src/main/java/at/core/ScreenShotListener.java
