+++ Core/src/main/java/at/core/ServiceHooks.java
