--- src/test/java/app/objects/Account.java
@@ -1,14 +1,0 @@
-package app.objects;
-
-/**
- * @author tien.hoang
- */
-public class Account {
-    public String email;
-    public String password;
-
-    public Account(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-}
