--- test.md
@@ -1,2 +1,0 @@
-|---|---|
-|bla|ble|
